+++ Service/Address/src/main/java/com/hzgc/service/address/controller/FtpAddressController.java
@@ -6,7 +6,6 @@ import com.hzgc.service.address.service.FtpAddressService;
 import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.netflix.feign.FeignClient;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
